fix: verbose text in histogram
